Adding the visualiser project
This is a basic audio visualiser which has a 2d and 3d version. In pure JavaScript using the pizzicato library which makes use of the web audio api. This is not supported across all browsers (surprise surprise it is IE) but seems to have reasonable support across other browsers. More information about the projects is included in the READMEs dotted around in each folder