Adding terrain generation
This was a really old project so I've gone through and tried to comment the code as best I can but I can't attest to how accurate or helpful they will be. I'm hoping that I've got the links in the README correct this time so it wont require an update at a later date.